Added a habit as an optional parameter of `NewHabitDialog`
This changes some text displayed depending on whether this parameter supplies null or not.
If supplied a `Habit`, the widget goes to modification mode.
Fill the UI entries with data supplied by the habit object passed through the parameter.
Deletes the supplied habit object from the database if the user confirms to modify via pressing "Modifying this Habit".