WIP Added `upcomingHabit` getter for getting the upcoming habit
This initial implementation queries the habit with recurring time greater and closest to the current time.
A flaw is that the way it gets this habit means that it doesn't consider habits that will recur the next week (when current weekday X is greater than weekday X-Y that will happen the next week) is not considered a upcoming habit. This needs to be fixed.