Updated `upcomingHabit` getter on `DatabaseHelper`
Fixed bug that passes a row when its hour is equal to current hour but its minute is lower than current minute, which is not upcoming.
This is fixed by adding another select statement that gets rows with start hours equal to current hour and start minutes greater than current minute.
The old first select statement is changed so that it compares using `>` instead of `>=`.