WIP Initial implementation of in-app notifications
Uses `cron` to schedule a debugger print that notifies a habit is starting.
`newSchedule` method is created to wipe previous cron jobs and create a new one based on the `upcomingHabit` data.
Added a listener to sqlite updates on initState of `_MainAppState` to recreate the cron schedule once a new insert happens on `habit_recurrance` table.