ENDWIP Fixed implementation of cron scheduling that causes a race condition and multiple calls to cron
Instead of closing and recreating a `Cron` object, `_HomePageState` now instead saves a `ScheduledTask` returned by `cron`'s `schedule` method and cancels the saved `ScheduledTask` every time the `newSchedule` method is called.
A `StreamSubscription` is also used to add a 'buffer'. This makes it so that when `newSchedule` is called, it waits for a short time, and only then will its main action be executed. If `newSchedule` is called before that short time finishes, it resets that delay. This makes it so that if multiple calls to `newSchedule` happens too quickly, it only accepts the last call.