Overhaul state management in `new_habit_dialog.dart`
Instead of creating another data type to store recurrance pairs, it now uses the existing `Habit` data type.
References to the removed data types are updated to instead use the `Habit` data type.
The state stores a boolean that knows whether the widget is in creation mode or modification mode.
Added a listener to the habit text controller to update the stored `Habit` to a new one with the updated name.