Added a `ongoingHabit` getter on `database_helper.dart`
This can be used with a cron job that runs every minute to monitor ongoing habits. I do think there is a better way to do this by allowing the user to 'run' an upcoming habit and tracking the end of that habit via a cron job. This getter would be kept if ever it finds a use.