Updated streams on `database_helper.dart`
Added a static cron instance for scheduling the updating of streams based on the start and end of upcoming and ongoing habits.
Cached latest stream results, maybe to just return that if the parameters are the same to avoid redundant sql queries.
Added my stream setup for upcoming habits.
Update streams when an ongoing habit ends.
Update streams when an upcoming habit starts.
Removed unnecessary `await` on some stream updaters.